requirements.txt
greenlet==3.0.2
pyodbc==5.0.1
python-dotenv==1.0.0
SQLAlchemy==2.0.23
typing_extensions==4.9.0
